singularity: Free constructors + separate files.
